refactor: Consolidate API response and utility functions
- Remove redundant api-response.ts file
- Merge API response and error handling logic into api-utils.ts
- Update import paths and error response structure
- Standardize error handling with consistent response format
